refactor: 리뷰 타입 파일 리팩토링 및 타입 추가
