--- frontend/src/types/detailReviewData.ts
@@ -1,16 +1,0 @@
-export interface ReviewItem {
-  question: string;
-  answer: string;
-}
-
-export interface DetailReviewData {
-  reviewer: {
-    memberId: number;
-    name: string;
-  };
-  reviewGroup: {
-    groupId: number;
-    name: string;
-  };
-  contents: ReviewItem[];
-}
